When attempting to login with an empty PIN, pass to pkcs11 as NULL
Passing a PIN of NULL is used to indicate that, e.g., PIN
authentication happens outside of the pkcs11 process (through e.g. an
external PIN pad). Since the empty string is generally used as the
absence of a value in Go, convert that to NULL in C.

(I can't entirely discount the possibility that some random device
somewhere actually has a 0-length PIN, but it seems unlikely...)