Allocate embedded pointers on C heap
This commit changes cAttributeList and cMechanismList to allocate shadow
copies of each CK_ATTRIBUTE and CK_MECHANISM on the C heap instead of
referencing byte slices on the Go heap. This is in response to upcoming
changes in the Go 1.6 cgo contract, when it becomes illegal to pass a Go
pointer containing other Go pointers to cgo.