Fix compilation on Windows amd64 (#24)
Apparently it is enough to specify the struct members names instead of passing them in order.

I apologize in advance for not testing this change (and also not really knowing what I'm doing). It compiles but I was not able to run the tests on a windows system due to the missing libsofthsm.