Modifies the newly-added GCMParams to allow reading IV
After submitting #71 we found this little snippet from Amazon's CloudHSM
documentation:

```
Note

When performing AES-GCM encryption, the HSM ignores the initialization
vector (IV) in the request and uses an IV that it generates. The HSM
writes the generated IV to the memory reference pointed to by the pIV
element of the CK_GCM_PARAMS parameters structure that you supply.
```

This isn't, so far as we can tell, a requirement of the v2.4 standard,
but CloudHSM is already used quite a lot and that will only grow.

This changes the structure to allow pulling out the IV after an encrypt
operation, but with the caveat that manual freeing is necessary. We
didn't see a better way of handling this without significantly changing
the API of the actual PKCS11 calls as they're mapped into the library.
This does make an assumption that no library will free this memory for
you during its handling of the Encrypt call, but that would be behavior
at odds with how all other structures are handled, where the client code
performs the free.

The other way it might be handled would be to add EncryptInitGCM and
EncryptGCM but it feels more intrusive to create special purpose calls
for individual ciphers.