Deprecate use of boolean return value from FindObjects (#26) (#81)
The previous behavior is not consistent with the specification; the only
way to detect the end of the search is when no objects are returned. In
the future this field should be removed but in the meantime there is no
need to force a breaking change.

Also updates p11.Session.FindObjects to no longer use the deprecated
field, and to handle arbitrarily large result sets.